+++ fast64_internal/oot/cutscene/motion/importer/__init__.py
@@ -1,1 +1,1 @@
-from .functions import importCutsceneMotion
+from .functions import importCutsceneData
