+++ fast64_internal/oot/cutscene/motion/importer/functions.py
@@ -1,7 +1,8 @@
-from bpy.types import Context
-from .classes import OOTCutsceneMotionImport
+import bpy
+
+from .classes import OOTCSMotionImport
 
 
-def importCutsceneMotion(context: Context, filename: str):
-    importData = OOTCutsceneMotionImport(context)
-    return importData.importFromC(filename)
+def importCutsceneData(filePath: str):
+    csMotionImport = OOTCSMotionImport(filePath)
+    return csMotionImport.setCutsceneData(bpy.context.scene.ootCSNumber)
